// src/schemas.ts
export type JsonValue =
  string | number | boolean | null | JsonValue[] | { [key: string]: JsonValue };

export type FormatRepresentation =
  "normalized" | "concise" | "keyValues" | "simplified";

export type FormatTemporal = "temporalValues" | "aggregatedValues";

export type EntityTypeName = string | string[];

export type Scope = string | string[];
// ...
export type Geometry =
  | GeometryPoint
  | GeometryMultiPoint
  | GeometryPolygon
  | GeometryLineString
  | GeometryMultiLineString
  | GeometryMultiPolygon;
// ...
export interface DateTimeValue {
  "@type": "DateTime";
  "@value": string;
}

export type EndpointAccept =
  "application/json" | "application/ld+json" | "application/geo+json";

export interface Endpoint {
  uri: string;
  accept?: EndpointAccept;
  timeout?: number;
  cooldown?: number;
  receiverInfo?: KeyValuePair[];
  notifierInfo?: KeyValuePair[];
}

export interface GeoProperty {
  type: "GeoProperty";
  value: Geometry;
  observedAt?: string;
  datasetId?: string;
  readonly createdAt?: string;
  readonly modifiedAt?: string;
  readonly deletedAt?: string;
  readonly instanceId?: string;
  readonly previousValue?: Geometry;
}

export interface Entity {
  id: string;
  type: EntityTypeName;
  scope?: Scope;
  location?: GeoProperty;
  observationSpace?: GeoProperty;
  operationSpace?: GeoProperty;
  readonly createdAt?: string;
  readonly modifiedAt?: string;
  readonly deletedAt?: string;
}
// ...
export interface Property<TValue = DateTimeValue | JsonValue> {
  type: "Property";
  value: TValue;
  observedAt?: string;
  unitCode?: string;
  datasetId?: string;
  readonly createdAt?: string;
  readonly modifiedAt?: string;
  readonly deletedAt?: string;
  readonly instanceId?: string;
  readonly previousValue?: TValue;
}
// ...
export interface Relationship<TTarget extends Entity = Entity> {
  type: "Relationship";
  object?: TTarget["id"] | TTarget["id"][];
  objectType?: TTarget["type"] | TTarget["type"][];
  observedAt?: string;
  datasetId?: string;
  readonly createdAt?: string;
  readonly modifiedAt?: string;
  readonly deletedAt?: string;
  readonly instanceId?: string;
  readonly previousObject?: TTarget["id"] | TTarget["id"][];
  readonly entity?: TTarget | TTarget[];
}